[MAIN] Add proxy memory update

+++ Cloud-Main/src/main/kotlin/dev/jonaz/cloud/model/docker/DockerInspectHostConfigModel.kt
@@ -70,7 +70,7 @@ data class DockerInspectHostConfigModel(
     @SerializedName("CpuShares")
     val cpuShares: Int,
     @SerializedName("Memory")
-    val memory: Int,
+    val memory: Long,
     @SerializedName("NanoCpus")
     val nanoCpus: Int,
     @SerializedName("CgroupParent")
